implementing code splitting and import statements for cleaner project structure
